test(logic): split quick battle resolver tests under test/combat
Move the former monolithic quick_battle_resolver_test.dart into three
files aligned with resolve vs build/siege vs apply/Spy concerns, each
under the 400-line logic test gate. Refs #2288.
